+++ Changelog.md
@@ -1,13 +1,26 @@
 # Change Log
+
 All major changes are listed here.
 
-# Feb 22nd, 2022
+# Jan 23rd, 2023
+
 ## Added
+
+- [x] Ubuntu20.04 with cuda11.3, cudnn8, python3.10
+- [x] Install python3.10 from source
+
+## Changed
+
+# Feb 22nd, 2022
+
+## Added
+
 - [x] Install `zsh` on `my_docker` image.
 - [x] Install Opencv4 on the dev docker image
 - [x] Install Pytorch1.9 on the dev docker image
 
 ## Changed
+
 - [x] Install python3.7 from source for base docker image
 - [x] Modify OpenCV installation scripts to make it work regardless of where python is installed for python3.7
 - [x] Modify OpenCV installation scripts to make it work regardless of where python is installed for python3.6